LayoutTests/fast/text/font-loading-multiple-sets-expected.txt

PASS font.status is "unloaded"
PASS set1.status is "loaded"
PASS set2.status is "loaded"
PASS font.status is "loading"
PASS set1.status is "loading"
PASS set2.status is "loading"
PASS successfullyParsed is true

TEST COMPLETE
